Support parsing of more fetch() errors
