+++ .changeset/five-emus-retire.md
@@ -1,0 +1,5 @@
+---
+"@atproto-labs/fetch": patch
+---
+
+Support parsing of more fetch() errors
