+++ packages/internal/fetch/src/fetch-request.ts
@@ -9,13 +9,85 @@ export class FetchRequestError extends FetchError {
     message?: string,
     options?: ErrorOptions,
   ) {
+    if (statusCode == null || !message) {
+      const info = extractInfo(extractRootCause(options?.cause))
+      statusCode ??= info[0]
+      message ||= info[1]
+    }
+
     super(statusCode, message, options)
+  }
+
+  get expose() {
+    // A 500 request error means that the request was not made due to an infra,
+    // programming or server side issue. The message should no be exposed to
+    // downstream clients.
+    return this.statusCode !== 500
   }
 
   static from(request: Request, cause: unknown): FetchRequestError {
     if (cause instanceof FetchRequestError) return cause
     return new FetchRequestError(request, undefined, undefined, { cause })
   }
+}
+
+function extractRootCause(err: unknown): unknown {
+  // Unwrap the Network error from undici (i.e. Node's internal fetch() implementation)
+  // https://github.com/nodejs/undici/blob/3274c975947ce11a08508743df026f73598bfead/lib/web/fetch/index.js#L223-L228
+  if (
+    err instanceof TypeError &&
+    err.message === 'fetch failed' &&
+    err.cause !== undefined
+  ) {
+    return err.cause
+  }
+
+  return err
+}
+
+function extractInfo(err: unknown): [statusCode: number, message: string] {
+  if (typeof err === 'string' && err.length > 0) {
+    return [500, err]
+  }
+
+  if (!(err instanceof Error)) {
+    return [500, 'Failed to fetch']
+  }
+
+  // Undici fetch() "network" errors
+  switch (err.message) {
+    case 'failed to fetch the data URL':
+      return [400, err.message]
+    case 'unexpected redirect':
+    case 'cors failure':
+    case 'blocked':
+    case 'proxy authentication required':
+      // These cases could be represented either as a 4xx user error (invalid
+      // URL provided), or as a 5xx server error (server didn't behave as
+      // expected).
+      return [502, err.message]
+  }
+
+  // NodeJS errors
+  const code = err['code']
+  if (typeof code === 'string') {
+    switch (true) {
+      case code === 'ENOTFOUND':
+        return [400, 'Invalid hostname']
+      case code === 'ECONNREFUSED':
+        return [502, 'Connection refused']
+      case code === 'DEPTH_ZERO_SELF_SIGNED_CERT':
+        return [502, 'Self-signed certificate']
+      case code.startsWith('ERR_TLS'):
+        return [502, 'TLS error']
+      case code.startsWith('ECONN'):
+        return [502, 'Connection error']
+      default:
+        return [500, `${code} error`]
+    }
+  }
+
+  return [500, err.message]
 }
 
 export function protocolCheckRequestTransform(protocols: {
